chore: move attributes to people page
