+++ apps/web/app/(app)/environments/[environmentId]/(peopleAndSegments)/attributes/components/AttributeSettingsTab.tsx
