+++ apps/web/app/(app)/environments/[environmentId]/(peopleAndSegments)/attributes/loading.tsx
