+++ apps/web/app/(app)/environments/[environmentId]/actions/components/ActionClassesTable.tsx
