+++ apps/web/app/(app)/environments/[environmentId]/actions/components/ActionDetailModal.tsx
