+++ apps/web/app/(app)/environments/[environmentId]/actions/layout.tsx
@@ -1,0 +1,9 @@
+import { ContentWrapper } from "@formbricks/ui/ContentWrapper";
+
+export default function ActionsAndAttributesLayout({ children }) {
+  return (
+    <>
+      <ContentWrapper>{children}</ContentWrapper>
+    </>
+  );
+}
